fix intensity param u16->u8

+++ programs/drift/src/instructions/admin.rs
@@ -515,8 +515,8 @@ pub fn handle_initialize_perp_market(
     order_tick_size: u64,
     min_order_size: u64,
     concentration_coef_scale: u128,
-    curve_update_intensity: u16,
-    amm_jit_intensity: u16,
+    curve_update_intensity: u8,
+    amm_jit_intensity: u8,
     name: [u8; 32],
 ) -> Result<()> {
     let perp_market_pubkey = ctx.accounts.perp_market.to_account_info().key;
